requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
pymongo
boto3
awscli